Integrate unknown pattern annotation tests into existing coverage
